requirements.txt
pyspark==3.3.0
kafka-python==2.0.2
psycopg2-binary===2.9.3
cassandra-driver==3.25.0


